Struct futures_spawn::NewThread

pub struct NewThread;

Spawn all futures on a new thread

This is the most basic Spawn implementation. Each call to spawn results in a new thread dedicated to processing the given future to completion.